home/ubuntu/video_generator_app/requirements.txt
# 必須ライブラリ
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.4.2
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6

# コメント抽出関連
tweepy==4.14.0

# 音声合成関連
openai==1.3.5
fish-audio-api==0.1.0

# 画像検索関連
apify-client==1.0.0
pillow==10.0.1

# 動画生成関連
moviepy==1.0.3
opencv-python==4.8.1.78
diffusionstudio-core==0.1.0
numpy==1.24.3

# WEBインターフェース関連
websockets==11.0.3
